refactor: combine motif files
also moved FillRect related code to rect.go
