+++ src/resources/defaultMotif.ini
@@ -6,12 +6,12 @@
 	localcoord = 320, 240
 
 [Files]
-	spr = data/system.sff
-	snd = data/system.snd
+	spr = system.sff
+	snd = system.snd
 	logo.storyboard = 
 	intro.storyboard = 
-	select = data/select.def
-	fight = data/fight.def
+	select = select.def
+	fight = fight.def
 	font0 = Open_Sans.def
 	font1 = default-3x5.def
 	font2 = f-6x9.def
@@ -32,7 +32,7 @@
 	font7.height = -1
 	font8.height = -1
 	font9.height = -1
-	glyphs = data/glyphs.sff
+	glyphs = glyphs.sff
 	module = 
 
 [Languages]
@@ -144,7 +144,7 @@
 	menu.tween.factor = 0.5
 	menu.tween.wrap.snap = 0
 
-	menu.item.font = f-6x9.def, 0, 1, 255, 255, 255, 255, -1
+	menu.item.font = f-6x9.def, 0, 1, 191, 191, 191, 255, -1
 	menu.item.offset = 0, 0
 	menu.item.scale = 1.0, 1.0
 	menu.item.xshear = 0
@@ -179,7 +179,7 @@
 	;menu.item.bg.<itemname>.window = 
 	;menu.item.bg.<itemname>.localcoord = 320, 240
 
-	menu.item.active.font = -1, 0, 0, 255, 255, 255, 255, -1
+	menu.item.active.font = f-6x9.def, 0, 0, 255, 255, 255, 255, -1
 	menu.item.active.offset = 0, 0
 	menu.item.active.scale = 1.0, 1.0
 	menu.item.active.xshear = 0
@@ -616,7 +616,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	p1.face2.spacing = 0, 0
 	p1.face2.padding = 0
 
-	p1.name.font = -1, 4, 1, 255, 255, 255, 255, -1
+	p1.name.font = jg.fnt, 4, 1, 255, 255, 255, 255, -1
 	p1.name.offset = 0, 0
 	p1.name.scale = 1.0, 1.0
 	p1.name.xshear = 0
@@ -757,7 +757,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 
 	p1.teammenu.value.spacing = 6, 0
 
-	p1.teammenu.item.font = -1, 0, 1, 255, 255, 255, 255, -1
+	p1.teammenu.item.font = f-6x9.def, 0, 1, 255, 255, 255, 255, -1
 	p1.teammenu.item.offset = 0, 0
 	p1.teammenu.item.scale = 1.0, 1.0
 	p1.teammenu.item.xshear = 0
@@ -1053,7 +1053,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	p2.face2.spacing = 0, 0
 	p2.face2.padding = 0
 
-	p2.name.font = -1, 1, -1, 255, 255, 255, 255, -1
+	p2.name.font = jg.fnt, 1, -1, 255, 255, 255, 255, -1
 	p2.name.offset = 0, 0
 	p2.name.scale = 1.0, 1.0
 	p2.name.xshear = 0
@@ -1190,7 +1190,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 
 	p2.teammenu.value.spacing = -6, 0
 
-	p2.teammenu.item.font = -1, 0, -1, 255, 255, 255, 255, -1
+	p2.teammenu.item.font = f-6x9.def, 0, -1, 255, 255, 255, 255, -1
 	p2.teammenu.item.offset = 0, 0
 	p2.teammenu.item.scale = 1.0, 1.0
 	p2.teammenu.item.xshear = 0
@@ -1590,7 +1590,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	p1.face2.spacing = 0, 0
 	p1.face2.padding = 0
 
-	p1.name.font = -1, 0, 0, 255, 255, 255, 255, -1
+	p1.name.font = jg.fnt, 0, 0, 255, 255, 255, 255, -1
 	p1.name.offset = 0, 0
 	p1.name.scale = 1.0, 1.0
 	p1.name.xshear = 0
@@ -1655,7 +1655,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	p2.anim = -1
 	p2.spr = 9000, 1
 	p2.offset = 0, 0
-	p2.facing = 1
+	p2.facing = -1
 	p2.scale = 1.0, 1.0
 	p2.xshear = 0
 	p2.angle = 0
@@ -1667,7 +1667,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	p2.done.anim = -1
 	p2.done.spr = 9000, 1
 	p2.done.offset = 0, 0
-	p2.done.facing = 1
+	p2.done.facing = -1
 	p2.done.scale = 1.0, 1.0
 	p2.done.xshear = 0
 	p2.done.angle = 0
@@ -1689,7 +1689,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	p2.face2.anim = -1
 	p2.face2.spr = -1, 0
 	p2.face2.offset = 0, 0
-	p2.face2.facing = 1
+	p2.face2.facing = -1
 	p2.face2.scale = 1.0, 1.0
 	p2.face2.xshear = 0
 	p2.face2.angle = 0
@@ -1701,7 +1701,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	p2.face2.done.anim = -1
 	p2.face2.done.spr = -1, 0
 	p2.face2.done.offset = 0, 0
-	p2.face2.done.facing = 1
+	p2.face2.done.facing = -1
 	p2.face2.done.scale = 1.0, 1.0
 	p2.face2.done.xshear = 0
 	p2.face2.done.angle = 0
@@ -1718,7 +1718,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	p2.face2.spacing = 0, 0
 	p2.face2.padding = 0
 
-	p2.name.font = -1, 0, 0, 255, 255, 255, 255, -1
+	p2.name.font = jg.fnt, 0, 0, 255, 255, 255, 255, -1
 	p2.name.offset = 0, 0
 	p2.name.scale = 1.0, 1.0
 	p2.name.xshear = 0
@@ -1865,6 +1865,8 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	select.enabled = 0
 	vsscreen.enabled = 0
 
+	cancel.key = a, b, c, x, y, z, s, m
+
 [Continue Screen]
 	enabled = 1
 
@@ -2010,6 +2012,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	enabled = 0
 	sounds.enabled = 0
 	cpu.enabled = 1
+	vs.enabled = 1
 	winner.teamko.enabled = 0
 
 	fadein.time = 0
@@ -2102,7 +2105,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	p2.anim = -1
 	p2.spr = 9000, 2
 	p2.offset = 0, 0
-	p2.facing = 1
+	p2.facing = -1
 	p2.scale = 1.0, 1.0
 	p2.xshear = 0
 	p2.angle = 0
@@ -2122,7 +2125,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	p2.face2.anim = -1
 	p2.face2.spr = -1, 0
 	p2.face2.offset = 0, 0
-	p2.face2.facing = 1
+	p2.face2.facing = -1
 	p2.face2.scale = 1.0, 1.0
 	p2.face2.xshear = 0
 	p2.face2.angle = 0
@@ -2958,7 +2961,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	menu.tween.factor = 0.5
 	menu.tween.wrap.snap = 0
 
-	menu.item.font = f-6x9.def, 0, 1, 255, 255, 255, 255, -1
+	menu.item.font = f-6x9.def, 0, 0, 191, 191, 191, 255, -1
 	menu.item.offset = 0, 0
 	menu.item.scale = 1.0, 1.0
 	menu.item.xshear = 0
@@ -2993,7 +2996,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	;menu.item.bg.<itemname>.window = 
 	;menu.item.bg.<itemname>.localcoord = 320, 240
 
-	menu.item.active.font = f-6x9.def, 0, 1, 255, 255, 255, 255, -1
+	menu.item.active.font = f-6x9.def, 0, 0, 255, 255, 255, 255, -1
 	menu.item.active.offset = 0, 0
 	menu.item.active.scale = 1.0, 1.0
 	menu.item.active.xshear = 0
@@ -3718,7 +3721,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	menu.tween.factor = 0.5
 	menu.tween.wrap.snap = 0
 
-	menu.item.font = -1, 0, 0, 255, 255, 255, 255, -1
+	menu.item.font = f-6x9.def, 0, 0, 191, 191, 191, 255, -1
 	menu.item.offset = 0, 0
 	menu.item.scale = 1.0, 1.0
 	menu.item.xshear = 0
@@ -3753,7 +3756,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	;menu.item.bg.<itemname>.window = 
 	;menu.item.bg.<itemname>.localcoord = 320, 240
 
-	menu.item.active.font = -1, 0, 0, 255, 255, 255, 255, -1
+	menu.item.active.font = f-6x9.def, 0, 0, 255, 255, 255, 255, -1
 	menu.item.active.offset = 0, 0
 	menu.item.active.scale = 1.0, 1.0
 	menu.item.active.xshear = 0
@@ -4001,7 +4004,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	p1.face.window = 
 	p1.face.localcoord = 320, 240
 
-	p1.name.font = -1, 0, 1, 255, 255, 255, 255, -1
+	p1.name.font = f-6x9.def, 0, 1, 255, 255, 255, 255, -1
 	p1.name.offset = 0, 0
 	p1.name.scale = 1.0, 1.0
 	p1.name.xshear = 0
@@ -4050,7 +4053,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	p2.face.anim = -1
 	p2.face.spr = 9000, 0
 	p2.face.offset = 0, 0
-	p2.face.facing = 1
+	p2.face.facing = -1
 	p2.face.scale = 1.0, 1.0
 	p2.face.xshear = 0
 	p2.face.angle = 0
@@ -4058,7 +4061,7 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	p2.face.window = 
 	p2.face.localcoord = 320, 240
 
-	p2.name.font = -1, 0, 1, 255, 255, 255, 255, -1
+	p2.name.font = f-6x9.def, 0, 1, 255, 255, 255, 255, -1
 	p2.name.offset = 0, 0
 	p2.name.scale = 1.0, 1.0
 	p2.name.xshear = 0
@@ -4300,10 +4303,11 @@ Welcome to I.K.E.M.E.N GO engine! (%s - %s)
 	next.key = $D, LS_Y+
 	previous.key = $U, LS_Y-
 	done.key = a, b, c, x, y, z
+	cancel.key = m
 
-	move.snd = 100, 0
-	cancel.snd = 100, 2
 	done.snd = 100, 1
+	cancel.snd = 100, 2
+	move.snd = 100, 0
 
 	done.time = 30
